\name{df_intersect}
\alias{df_intersect}
\docType{data}
\title{
	Example data for plotting output of samExplore function.
}
\description{
	Example data for plotting output of samExplore function.
}
\usage{data("df_intersect")}
\format{
  A data frame with 1125 observations on the following 3 variables.
  \describe{
    \item{\code{Label}}{a character vector}
    \item{\code{Variable}}{a numeric vector}
    \item{\code{Value}}{a numeric vector}
  }
}
\details{
 is a dataframe object
}
\value{}
\examples{
data(df_intersect)

}
\keyword{datasets}
