\name{df_intersect}
\alias{df_intersect}
\docType{data}
\title{
    Example data for plotting output of samExplore function.
}
\description{
    Example data for plotting output of samExplore function. Dataframe consists three columns,
first column contains names of new Genes or exons, second column provides sequence depth 
and third column provide total counts for the corresponding sequence coverage.
}
\usage{data("df_intersect")}
\format{
    A data frame with 1125 observations on the following 3 variables.
      \describe{
        \item{\code{Label}}{a character vector}
        \item{\code{Variable}}{a numeric vector}
        \item{\code{Value}}{a numeric vector}
    }
}
\details{
    Example data for plotting output of samExplore function. Dataframe consists three columns,
first column contains names of new Genes or exons, second column provides sequence depth 
and third column provide total counts for the corresponding sequence coverage.
}
\value{Example data for plotting results.}
\examples{
data(df_intersect)

}
\keyword{datasets}
